requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==23.2.0
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
hydra-core==1.3.2
idna==3.7
Jinja2==3.1.3
kiwisolver==1.4.5
lightning==2.2.2
lightning-utilities==0.11.2
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.2
PyYAML==6.0.1
setuptools==68.2.2
six==1.16.0
sympy==1.12
torch==2.2.2
torchmetrics==1.3.2
tqdm==4.66.2
typing_extensions==4.11.0
wheel==0.41.2
yarl==1.9.4
